// generated/src/app/api-models/commission-determ-x.model.ts
/* Copyright (c) 2020 . All Rights Reserved. */

export class CommissionDetermX {
  matrixDeterminant: string;
  matrixSeq: number;
  xAxisDescription: string;
  livesFrom: number;
  livesThru: number;
  casesFrom: number;
  casesThru: number;
  maximumLivesForACase: number;
  agentType: string;
  lineOfBusiness: string;
  yearsLicensedFrom: number;
  yearsLicensedThru: number;
  yearsPeriod: string;
  securityCode: string;
  insertDatetime: Date;
  insertUser: string;
  insertProcess: string;
  updateDatetime: Date;
  updateUser: string;
  updateProcess: string;
}